refactor: consolidate redux slices
