+++ hooks/use-session-state.ts
@@ -5,7 +5,7 @@ import {
   setDomain,
   setTenantId,
   resetAuthState,
-} from "@/lib/redux/slices/app-config";
+} from "@/lib/redux/slices/app-state";
 import { ErrorManager } from "@/lib/error-handling/error-manager";
 
 interface SessionState {
@@ -18,7 +18,7 @@ interface SessionState {
 export function useSessionState(): SessionState {
   const { data: session, status, update } = useSession();
   const dispatch = useAppDispatch();
-  const appConfig = useAppSelector((state) => state.appConfig);
+  const appConfig = useAppSelector((state) => state.appState);
   const lastValidation = useRef<number>(Date.now());
 
   useEffect(() => {
