+++ hooks/use-step-execution.ts
@@ -1,11 +1,11 @@
 import { executeStepAction } from "@/app/actions/step-actions";
 import { ErrorManager } from "@/lib/error-handling/error-manager";
-import { addOutputs } from "@/lib/redux/slices/app-config";
-import { addApiLog } from "@/lib/redux/slices/debug-panel";
+import { addOutputs } from "@/lib/redux/slices/app-state";
+import { addApiLog } from "@/lib/redux/slices/ui-state";
 import {
   clearCheckTimestamp,
   updateStep,
-} from "@/lib/redux/slices/setup-steps";
+} from "@/lib/redux/slices/app-state";
 import { allStepDefinitions } from "@/lib/steps";
 import type { StepId } from "@/lib/steps/step-refs";
 import { useCallback } from "react";
@@ -13,7 +13,7 @@ import { useAppDispatch, useAppSelector } from "./use-redux";
 
 export function useStepExecution() {
   const dispatch = useAppDispatch();
-  const appConfig = useAppSelector((state) => state.appConfig);
+  const appConfig = useAppSelector((state) => state.appState);
 
   const executeStep = useCallback(
     async (stepId: StepId) => {
