+++ lib/redux/store.ts
@@ -1,17 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { appConfigSlice } from "./slices/app-config";
-import { setupStepsSlice } from "./slices/setup-steps";
-import { modalsSlice } from "./slices/modals";
-import { errorsSlice } from "./slices/errors";
-import { debugPanelSlice } from "./slices/debug-panel";
+import { appStateSlice } from "./slices/app-state";
+import { uiStateSlice } from "./slices/ui-state";
 
 export const store = configureStore({
   reducer: {
-    appConfig: appConfigSlice.reducer,
-    setupSteps: setupStepsSlice.reducer,
-    modals: modalsSlice.reducer,
-    errors: errorsSlice.reducer,
-    debugPanel: debugPanelSlice.reducer,
+    appState: appStateSlice.reducer,
+    uiState: uiStateSlice.reducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
